+++ src/cache.cc
@@ -37,6 +37,20 @@
 
 using namespace std;
 
+const char *CACHE_VERSION = "1";
+
+static std::ostream& operator<<(std::ostream& out, const CachePage& page) {
+	out << page.label << '\0';
+	out << page.text << '\0';
+	return out;
+}
+
+static std::istream& operator>>(std::istream& in, CachePage& page) {
+	std::getline(in, page.label, '\0');
+	std::getline(in, page.text, '\0');
+	return in;
+}
+
 Cache::Cache(string const &cache_file)
 	: cache_file(cache_file), valid(false) {
 	// Open the cache file
@@ -47,26 +61,32 @@ Cache::Cache(string const &cache_file)
 
 	unsigned char indicator;
 	fd >> indicator;
-	if (indicator != 'G') {
+	if (indicator != 'C') {
 		return;
 	}
-	for (std::string page; std::getline(fd, page, '\0'); ) {
+	std::string version;
+	std::getline(fd, version, '\0');
+	if (version != CACHE_VERSION) {
+		return;
+	}
+
+	for (CachePage page; fd >> page; ) {
 		pages.push_back(page);
 	}
 	valid = true;
 }
 
-void Cache::set_page(unsigned pagenum, const string &text) {
+void Cache::set_page(unsigned pagenum, const CachePage& page) {
 	pages.resize(max(pagenum, (unsigned)pages.size()));
-	pages[pagenum-1] = text;
+	pages[pagenum-1] = page;
 }
 
-bool Cache::get_page(unsigned pagenum, string &text) {
+bool Cache::get_page(unsigned pagenum, CachePage& page) {
 	if (!valid) {
 		return false;
 	}
 	if (pagenum-1 < pages.size()) {
-		text = pages[pagenum-1];
+		page = pages[pagenum-1];
 		return true;
 	}
 	return false;
@@ -77,18 +97,19 @@ void Cache::dump() {
 	if (!fd) {
 		return;
 	}
-	// The first byte of the cache file is an indicator byte, which
-	// can be written atomically. Initial it is \0, after all pages
-	// have been flushed, the indicator byte is written to 'G' (for
-	// good).
+	// The first byte of the cache file is an indicator byte, which can be
+	// written atomically. Initial it is \0, after all pages have been
+	// flushed, the indicator byte is written to 'C' (complete cache).
 	fd << '\0';
-	for (const string &page : pages) {
+	fd << CACHE_VERSION << '\0';
+
+	for (const CachePage& page : pages) {
 		fd << page;
-		fd << '\0';
 	}
+
 	fd.flush();
 	fd.seekp(0, ios_base::beg);
-	fd << 'G';
+	fd << 'C';
 	fd.close();
 }