+++ src/output.h
@@ -26,10 +26,11 @@
 #include "pdfgrep.h"
 
 struct context {
-	const std::string &filename;
+	const std::string& filename;
 	size_t pagenum;
+	std::string page_label;
 
-	const Outconf &out;
+	const Outconf& out;
 };
 
 struct match {
@@ -65,10 +66,7 @@ void print_context_separator(const Outconf &out);
 // Print to stderr, with "pdfgrep: " as prefix;
 std::ostream& err();
 
-std::ostream& line_prefix(const Outconf& outconf, const std::string& filename,
-                          bool in_context, size_t page);
-std::ostream& line_prefix(const Outconf& outconf, const std::string& filename,
-                          bool in_context);
+std::ostream& line_prefix(const context& context, bool in_context);
 
 #endif